requirements-dev.txt
-r requirements.txt
pytest-factoryboy==2.1
pytest-django==4.2
pdoc
django-extensions
pre-commit
